// api/src/__generated__/accounts.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: 1JLjgdrTGcvrVQu/tP0L6Ocb9DNcMvvlYqSFz6f1bKB8iGrDK+OHEm3uHRHCtiQSw6jlna6MJJiSts4yOcBM7g==
 */

// eslint:disable
// tslint:disable

interface Accounts {
  /**
   * @default false
   */
  active: boolean
  balance: number
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at: (Date) | null
  /**
   * @default nextval('accounts_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'accounts_id'}
  inactive_since: (Date) | null
  name: string
  password: string
  /**
   * @default true
   */
  print: boolean
  uid: string
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at: (Date) | null
}
export default Accounts;

interface Accounts_InsertParameters {
  /**
   * @default false
   */
  active?: boolean
  balance: number
  /**
   * @default CURRENT_TIMESTAMP
   */
  created_at?: (Date) | null
  /**
   * @default nextval('accounts_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'accounts_id'}
  inactive_since?: (Date) | null
  name: string
  password: string
  /**
   * @default true
   */
  print?: boolean
  uid: string
  /**
   * @default CURRENT_TIMESTAMP
   */
  updated_at?: (Date) | null
}
export type {Accounts_InsertParameters}